back/requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
Flask-Migrate
Flask-Script
marshmallow
